Reorganised Test Files + fixed invocation bug
