--- src/main/java/com/SENG4430/ClassCoupling/CouplingMetricCalculator.java
@@ -1,17 +1,0 @@
-package com.SENG4430.ClassCoupling;
-
-import spoon.Launcher;
-import spoon.reflect.CtModel;
-import spoon.reflect.code.*;
-import spoon.reflect.declaration.*;
-import spoon.reflect.visitor.filter.TypeFilter;
-
-import java.util.HashMap;
-
-public class CouplingMetricCalculator {
-
-    public static int calculateClassCoupling(String inputFilePath) {
-        return 0;
-    }
-
-}
