+++ src/main/java/com/SENG4430/ClassCoupling/TestFiles/BadCode.java
@@ -1,4 +1,4 @@
-package com.SENG4430.ClassCoupling;
+package com.SENG4430.ClassCoupling.TestFiles;
 
 public class BadCode {
     public BadCode()
