+++ src/main/java/com/SENG4430/ClassCoupling/TestFiles/CouplingMetricCalculator.java
@@ -1,0 +1,9 @@
+package com.SENG4430.ClassCoupling.TestFiles;
+
+public class CouplingMetricCalculator {
+
+    public static int calculateClassCoupling(String inputFilePath) {
+        return 0;
+    }
+
+}
